Validate presence and uniqueness of event name and year
It seems sensible that the event name should be unique for that
particular year, i.e. the overall string representation of the event
(“RubyConf 2017”) identifies it uniquely.

It may not be true that all events necessarily have a year in their
name, but the app already makes that assumption so we may as well codify
it in the validations. (If the assumption changes, the validations will
have to change too.)